CR Change: Move StreamHelper to impl package and improve javadoc.
