+++ client/src/main/java/io/pravega/client/admin/impl/StreamCutHelper.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.client;
+package io.pravega.client.admin.impl;
 
 import io.pravega.client.netty.impl.ConnectionFactory;
 import io.pravega.client.segment.impl.Segment;
