+++ client/src/main/java/io/pravega/client/admin/impl/StreamManagerImpl.java
@@ -12,7 +12,6 @@ package io.pravega.client.admin.impl;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import io.pravega.client.ClientConfig;
-import io.pravega.client.StreamCutHelper;
 import io.pravega.client.admin.StreamInfo;
 import io.pravega.client.admin.StreamManager;
 import io.pravega.client.netty.impl.ConnectionFactory;
